Big protocol refactoring
Adjust the protocol to just one service; All item types are no simply
addressed by a u64. The device item no longer proxies for the other
types but simply exposes their index if a matching one is found, leaving
it to the client to resolve the underlying items and call functions
directly on them.