cli: Improve help message of "configure" sub-command
boardswarm-cli is used to access any instance by associating an easy to
use name with the particular instance. In case the user runs both an
instance and boardswarm-cli on the same machine, the message "Configure
boardswarm instances" might be confusing, because boardswarm instances are
boardswarm servers, which are configured with their own server.conf. Make
it more explicit that with boardswarm-cli we configure client
authentication to boardswarm servers.

Signed-off-by: Andrzej Pietrasiewicz <andrzej.p@collabora.com>
Co-authored-by: Sjoerd Simons <sjoerd@luon.net>