Wrap ItemType to use #[arg(value_enum)]
This way we plug into Clap's help generation system.
We cannot simply apply #[arg(value_enum)] to ItemType, because the enum
uses the nouns in singular, whereas we want the commands to look
e.g. like this:

$ boardswarm-cli list devices

rather than

$ boardswarm-cli list device

We cannot explicitly reimplement ValueEnum for ItemType, because both
the trait and the enum are foreign, so we wrap ItemType into our
local newtype struct and then explicitly implement ValueEnum for it.
This way we can customize how user input is converted to actual values.

Signed-off-by: Andrzej Pietrasiewicz <andrzej.p@collabora.com>
Co-authored-by: Sjoerd Simons <sjoerd@luon.net>