boardswarm: Tweak example server.conf to just work
In it's current state, the current example server.conf file defaults to
enabling a number of options that either require extra external
configuration or specific hardware to "just work". This makes it harder
to get to a point where the server runs and basic operation can be
tested.

Block comment out configuration options that required more configuration
to be operational to make it easier to use as part of the basic
installation guide.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>