boardswarm: Use system user and group
The current example systemd unit starts the server using a dynamic
boardswarm user. Dynamic systemd users are created when the service
starts and and cleaned up when the service is terminated.

The side effect of this is that the boardswarm user and group aren't
available at boot when udev applies rules for already attached devices,
such as USB serial adapters. This results in udev being unable to set
the group as specified in the (example) udev rules and boardswarm being
unable to access the serial consoles.

The basic installation guide previously added documents adding a
boardswarm system user and group. Update the example systemd unit to use
these.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>