client: config: avoid extra copy when parsing token
Not that it really matters given the size of tokens, but because we can.
First deserialize to a &str, trim it and then make an owned copy of the
slice. Rather then asking serde to copy to an owned string first only to
drop it right after making another copy.