Distribute boardswarm binaries
To make it easier to test/install boardswarm also build binaries.
Unfortunately it's not easy to download assets from action runs. The
best way to do that seems to be to create a (pre)release and attach the
files there.

This builds on debian bookworm targetting both 64 bit ARM and Intel
systems.