client: Better track volume write/flush/shutdown errors
Not all users necessarily keep polling for flush or shutdown until
they're finished. In particular tokio::io::copy happens to do this as
tries to flush the output if there is no data on the input to avoid
deadlocks if there is an interdependency.

As the boardswarm server may only emit errors on e.g. flush, dropping
a flush results ends up masking errors. To avoid that don't just
monitor pending writes, but also pending flushes and shutdowns (iotw any
request that can modify the remote target). This allows bubbling the
error of any previous request as they come in.