client: Push responsibility to align to the blocksize to users
VolumeIoRW cannot differentiate between the write being not aligned to
the blocksize because the user is sloppy or because it's the final part
of a continous write. In the latter case the current behaviour would
trigger one big blocksize aligned write, followed by a small write with
the remaining data.

Instead expose the maximum accepted write size and move the alignment
responsibility to the user. This makes writing a bit more efficient for
good users.