cli: Extend batching support
Make the batching helper responsible for trying to match the volumes
blocksize as well as maximizing the amount sent per request.

Also make it possible to discard flush requests as they're not really
required for write-only usage and force a network roundtrips. This does
require the user to either call `shutdown` or `force_flush` before
dropping to ensure all data is written.