boardswarm: Add volume target field to device commit command
Android's fastboot tool support the `stage` and `continue` commands to
download a binary into memory a continue booting from there. This can be
implemented in the fastboot provider as a write to a specific `stage`
volume target, followed by a commit.

At the moment, boardswarm volume commit is only related to a device
volume, but unrelated to a specific target. In order to implement the
above, let's add another field for the volume target.

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@collabora.com>