boardswarm: Add eswin-eic7700-storage provider
The eswin-eic7700-storage provider registers a volume with a single
target `bootloader`.

When in USB download mode, the Hifive P550 will register itself as a
mass storage device. When the `bootloader_ddr5_secboot.bin` bootloader
artifact is written to its filesystem, it will automatically unregister
and bootstrap into this bootloader image.

Writing to the volume target `bootloader` will create and write the
content to this file `bootloader_ddr5_secboot.bin`.

Example server configuration:
```
providers:
  - name: eswin-eic7700-storage
    provider: eswin-eic7700-storage
    parameters:
      match:
        udev.ID_VENDOR: "ESWIN"
        udev.ID_MODEL: "WIN-2030"

devices:
  - name: sifive-hifive-p550-evk-cbg-0
    consoles:
    volumes:
      - name: storage
        match:
          udev.ID_PATH: "pci-0000:02:00.2-usb-0:1.1.2:1.0-scsi-0:0:0:0"
```

The user running boardswarm need to have access to the directory where
the storage device is mounted.

For example, a udev rule like the following could be added to automount
the device:
```
$ cat /etc/udev/rules.d/99-boardswarm.rules
ACTION=="add", SUBSYSTEMS=="usb", SUBSYSTEM=="block", \
    ENV{ID_VENDOR}=="ESWIN", ENV{ID_MODEL}=="WIN-2030", TAG+="systemd", \
    RUN{program}+="/usr/bin/systemd-mount --owner=boardswarm --no-block --automount=no --collect $devnode /var/lib/boardswarm/$env{ID_PATH}"
```

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@collabora.com>