boardswarm: udev: Refactor PreRegistration
In order to extend the device preregistration to other boardswarm items
(e.g. actuators) remove the requirement between Registrations and
DeviceRegistrations traits as that tights it to the Volume trait as
well.

We loose generality, but at the moment, there doesn't seem to be a
better way to implement this a sort of generic trait definition.

Signed-off-by: Ariel D'Alessandro <ariel.dalessandro@collabora.com>