Make registry indexes opaque
Rather then hardcoding u64 as the registry index, use a generic type
that implements the RegistryIndex trait. It's up to the registry index
implementor to make sure each index generated is unique and the index
space is big enough that it won't realistically overflow.

The required refactoring also makes the various users more generic as a
preperation for using unique index type per item.