Add GPIO control documentation
Some very basic documentation on how to drive GPIOs. This is certainly
not the most ideal way to drive a GPIO, however I currently haven't
found a better way to drive an arbitrary GPIO line (say one hooked up to
a user definable GPIO input). The syntax required for `<MODE>` wasn't
obvious as a user, hence documenting this for others.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>