Add page for Mediatek boards
This page currently covers the genio 700 EVK, but adding a more general
page so that it can be expanded easier in the future for more similar
boards.

Some of the very generic sections have been split out so that they can
be included into this page and other pages in the future. I feel that
having all the details on one page aids with following a guide,
especially when new to something, rather than needing to jump all over
the place to read sections on differnet pages.

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>